Files are now checked before download. The check ensures that the file does not already exist, and if it does exist, then the sizes of the remote and local files are compared. If they are the same size, the file is skipped. This ensures that complete files are not overwritten, and that new files can still be downloaded. eg. If you downloaded the first three pages of a chapter, and half of the fourth, you would only need to redownload the fourth.

+++ URLParser.py
@@ -220,8 +220,12 @@ class URLParser:
 	
 	def Download(self, url, workDir, statusbar):
 		filep = URLParser.LastFileInPath(self, url);
-		statusbar.SetStatusText('Downloading: ' + filep)
-		urllib.urlretrieve(url, workDir + "/" + filep)
+		lFile = workDir + "/" + filep
+		if os.path.isfile(lFile) and os.path.getsize(lFile) == int(urllib2.urlopen(url).headers["Content-Length"]):
+			statusbar.SetStatusText(filep + ' already exists')
+		else:
+			statusbar.SetStatusText('Downloading: ' + filep)
+			urllib.urlretrieve(url, workDir + "/" + filep)
 	
 	def LastFileInPath(self, path):
 		start = path.rindex('/')
